test ability change
tested the new flow of having predicted/server/sim ticks on abilities separated out, made the spell line trace on the client and then spawn a particle effect, passing the explosion vector to the server as a prediction param. server does a sphere check at the explosion locations and deals damage to actors, passes the locations as a broadcast param to sim clients, who also do the particle effect.