refactor of combat params
implemented new FCombatParameter struct that contains a pointer, location, rotation, and scale, along with an ID number and a ParamType enum.

implemented custom NetSerialize for FCombatParameter that should hopefully avoid serializing a Vector or Rotator if they are set to 0, thus saving bandwidth. I have no idea if that's actually how NetSerializeOptionalValue works, but it sounds nice!