Threat done?
- Scrapped combat groups. Everyone just binds to everything. This also means that things like linking, finding in-combat allies, and proximity aggro need to take place at the blueprint/ai level, and are not hard-coded into the threat system.
- Interactions that ARE hard-coded:
When a target is healed, the healer generates threat on any actors targeting the healed target.
When a targeted actor performs a heal on another target, that is NOT currently targeted, we add threat for the healer and add the healed target to the threat target.
Taking damage generates threat for the damager.
Dying clears the threat table, and also triggers delegates in targeting actors to remove the dead target from their own threat tables.
- Threat special moves should all be set up:
Taunt, Drop Threat, Transfer Threat, and Vanish for instant effects, Fixate, Blind, Fade, Misdirect for buff-based duration effects.
- Removed threat decay, because why?
- Removed some checks because I went a little crazy on them.